feat: DSTEW-1833 moved ReadModels to usecase package rather than domain
