+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/infrastructure/jpa/getapplication/GetApplicationGatewayMapper.java
@@ -7,17 +7,17 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 import tools.jackson.databind.ObjectMapper;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationProceedingReadModel;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationReadModel;
-import uk.gov.justice.laa.dstew.access.domain.InvolvedChildReadModel;
-import uk.gov.justice.laa.dstew.access.domain.OpponentReadModel;
-import uk.gov.justice.laa.dstew.access.domain.ProviderReadModel;
-import uk.gov.justice.laa.dstew.access.domain.ScopeLimitationReadModel;
 import uk.gov.justice.laa.dstew.access.entity.ApplicationEntity;
 import uk.gov.justice.laa.dstew.access.entity.ProceedingEntity;
 import uk.gov.justice.laa.dstew.access.model.InvolvedChild;
 import uk.gov.justice.laa.dstew.access.model.Opposable;
 import uk.gov.justice.laa.dstew.access.model.ProceedingLinkedChild;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationProceedingReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.InvolvedChildReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.OpponentReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ProviderReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ScopeLimitationReadModel;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.ApplicationContent;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.ApplicationMerits;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.Proceeding;
