+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/usecase/getapplication/GetApplicationUseCase.java
@@ -3,10 +3,10 @@ package uk.gov.justice.laa.dstew.access.usecase.getapplication;
 import jakarta.transaction.Transactional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationReadModel;
 import uk.gov.justice.laa.dstew.access.exception.ResourceNotFoundException;
 import uk.gov.justice.laa.dstew.access.security.AllowApiCaseworker;
 import uk.gov.justice.laa.dstew.access.usecase.getapplication.infrastructure.GetApplicationApplicationGateway;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationReadModel;
 
 /** Orchestrates retrieval of a single application. */
 @RequiredArgsConstructor
