+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/usecase/getapplication/infrastructure/GetApplicationApplicationGateway.java
@@ -2,7 +2,7 @@ package uk.gov.justice.laa.dstew.access.usecase.getapplication.infrastructure;
 
 import java.util.Optional;
 import java.util.UUID;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationReadModel;
 
 /** Gateway for loading a single application read model. */
 public interface GetApplicationApplicationGateway {
