+++ data-access-service/src/test/java/uk/gov/justice/laa/dstew/access/infrastructure/jpa/getapplication/GetApplicationGatewayMapperTest.java
@@ -11,8 +11,6 @@ import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tools.jackson.databind.ObjectMapper;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationProceedingReadModel;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationReadModel;
 import uk.gov.justice.laa.dstew.access.entity.ApplicationEntity;
 import uk.gov.justice.laa.dstew.access.entity.CaseworkerEntity;
 import uk.gov.justice.laa.dstew.access.entity.DecisionEntity;
@@ -21,6 +19,8 @@ import uk.gov.justice.laa.dstew.access.model.ApplicationStatus;
 import uk.gov.justice.laa.dstew.access.model.DecisionStatus;
 import uk.gov.justice.laa.dstew.access.model.OpponentDetails;
 import uk.gov.justice.laa.dstew.access.model.ProceedingLinkedChild;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationProceedingReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationReadModel;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.ApplicationContent;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.ApplicationMerits;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.ProceedingMerits;
