+++ data-access-service/src/test/java/uk/gov/justice/laa/dstew/access/usecase/getapplication/GetApplicationUseCaseTest.java
@@ -13,9 +13,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationReadModel;
 import uk.gov.justice.laa.dstew.access.exception.ResourceNotFoundException;
 import uk.gov.justice.laa.dstew.access.usecase.getapplication.infrastructure.GetApplicationApplicationGateway;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationReadModel;
 import uk.gov.justice.laa.dstew.access.utils.generator.DataGenerator;
 import uk.gov.justice.laa.dstew.access.utils.generator.getapplication.ApplicationReadModelGenerator;
