+++ data-access-service/src/testUtilities/java/uk/gov/justice/laa/dstew/access/utils/generator/getapplication/ApplicationReadModelGenerator.java
@@ -3,8 +3,8 @@ package uk.gov.justice.laa.dstew.access.utils.generator.getapplication;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
-import uk.gov.justice.laa.dstew.access.domain.ApplicationReadModel;
-import uk.gov.justice.laa.dstew.access.domain.ProviderReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ApplicationReadModel;
+import uk.gov.justice.laa.dstew.access.usecase.getapplication.model.ProviderReadModel;
 import uk.gov.justice.laa.dstew.access.utils.generator.BaseGenerator;
 
 /** Generator for {@link ApplicationReadModel} test data. */
